// apps/server-telemetry/src/app/modules/victoria-logs/types.ts
/*
 * Copyright (c) 2025.
 * Author Peter Placzek (tada5hi)
 * For the full copyright and license information,
 * view the LICENSE file that was distributed with this source code.
 */

export type VictoriaLogsModuleOptions = {
    baseURL?: string;
    ingestorURL?: string;
    querierURL?: string;
};